+++ Shared/CommandHandlers/SensorCommandHandler.cpp
@@ -11,7 +11,7 @@ SensorCommandHandler::SensorCommandHandler(BroadcastManager* broadcastManager, W
 {
 }
 
-bool SensorCommandHandler::handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], int paramCount)
+bool SensorCommandHandler::handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], uint8_t paramCount)
 {
     String cmd = command;
     cmd.trim();
