+++ Shared/CommandHandlers/SensorCommandHandler.h
@@ -33,7 +33,7 @@ public:
 	explicit SensorCommandHandler(BroadcastManager* broadcastManager, WarningManager* warningManager);
 #endif
 
-    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], int paramCount) override;
+    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], uint8_t paramCount) override;
     const String* supportedCommands(size_t& count) const override;
 
 	float getTemperature() const { return _lastTemperature; }
