+++ Shared/CommandHandlers/WarningCommandHandler.h
@@ -35,7 +35,7 @@ public:
 	explicit WarningCommandHandler(BroadcastManager* broadcastManager, WarningManager* warningManager);
 #endif
 
-    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], int paramCount) override;
+    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], uint8_t paramCount) override;
     const String* supportedCommands(size_t& count) const override;
 };
