+++ SmartFuseBox/ConfigCommandHandler.cpp
@@ -11,7 +11,7 @@ ConfigCommandHandler::ConfigCommandHandler(SoundManager* soundManager, Bluetooth
 {
 }
 
-bool ConfigCommandHandler::handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], int paramCount)
+bool ConfigCommandHandler::handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], uint8_t paramCount)
 {
     // Access the in-memory config
     Config* config = ConfigManager::getConfigPtr();
