+++ SmartFuseBox/SoundCommandHandler.h
@@ -15,7 +15,7 @@ private:
 
 public:
     SoundCommandHandler(SerialCommandManager* commandMgrComputer, SerialCommandManager* commandMgrLink, SoundManager* soundManager);
-    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], int paramCount) override;
+    bool handleCommand(SerialCommandManager* sender, const String command, const StringKeyValue params[], uint8_t paramCount) override;
 
     const String* supportedCommands(size_t& count) const override;
 private:
